requirements.txt
ultralytics>=8.0.0
onnxruntime
onnx
captcha_recognizer
fastapi[standard]
pillow
